@@ -1,8 +1,12 @@
 name: connectivity_plus
 description: Flutter plugin for discovering the state of the network (WiFi & mobile/cellular) connectivity on Android and iOS.
-version: 0.8.1
+version: 0.9.0-nullsafety.0
 homepage: https://plus.fluttercommunity.dev/
 repository: https://github.com/fluttercommunity/plus_plugins/tree/main/packages/
+
+environment:
+  sdk: ">=2.12.0-0 <3.0.0"
+  flutter: ">=1.12.13+hotfix.5"
 
 flutter:
   plugin:
@@ -24,23 +28,18 @@ flutter:
 dependencies:
   flutter:
     sdk: flutter
-  meta: ^1.0.5
-  connectivity_plus_platform_interface: ^0.4.1
-  connectivity_plus_linux: ^0.3.1
-  connectivity_plus_macos: ^0.4.1
-  connectivity_plus_web: ^0.6.1
-  connectivity_plus_windows: ^0.1.1
+  meta: ^1.3.0
+  connectivity_plus_platform_interface: ^0.5.0-nullsafety
+  connectivity_plus_linux: ^0.4.0-nullsafety
+  connectivity_plus_macos: ^0.5.0-nullsafety
+  connectivity_plus_web: ^0.7.0-nullsafety
+  connectivity_plus_windows: ^0.2.0-nullsafety
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_driver:
-    sdk: flutter
   test: any
-  mockito: ^4.1.1
-  plugin_platform_interface: ^1.0.0
-  pedantic: ^1.9.2
+  mockito: ^5.0.0-nullsafety.7
+  plugin_platform_interface: ^2.0.0
+  pedantic: ^1.10.0
 
-environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.12.13+hotfix.5"
