@@ -2,24 +2,25 @@ name: connectivity_plus_linux
 description: Linux implementation of the connectivity_plus plugin
 homepage: https://plus.fluttercommunity.dev/
 repository: https://github.com/fluttercommunity/plus_plugins/tree/main/packages/
-version: 0.3.1
+version: 0.4.0-nullsafety.1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
   flutter: ">=1.20.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  connectivity_plus_platform_interface: ^0.4.1
-  dbus: ^0.1.0
-  meta: ^1.2.3
+  connectivity_plus_platform_interface: ^0.5.0-nullsafety.0
+  dbus: ^0.2.0
+  meta: ^1.3.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mockito: ^4.1.2
-  pedantic: ^1.9.2
+  mockito: ^5.0.0-nullsafety.7
+  pedantic: ^1.10.0
+  build_runner: ^1.11.5
 
 flutter:
   plugin:
