@@ -1,8 +1,12 @@
 name: connectivity_plus_macos
 description: macOS implementation of the connectivity_plus plugin.
-version: 0.4.1
+version: 0.5.0-nullsafety.0
 homepage: https://plus.fluttercommunity.dev/
 repository: https://github.com/fluttercommunity/plus_plugins/tree/main/packages/
+
+environment:
+  sdk: ">=2.12.0-0 <3.0.0"
+  flutter: '>=1.20.0'
 
 flutter:
   plugin:
@@ -10,11 +14,7 @@ flutter:
       macos:
         pluginClass: ConnectivityPlugin
 
-environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: '>=1.20.0'
-
 dependencies:
-  connectivity_plus_platform_interface: ^0.4.1
+  connectivity_plus_platform_interface: ^0.5.0-nullsafety.0
   flutter:
     sdk: flutter
