@@ -2,10 +2,14 @@ import 'package:connectivity_plus_linux/src/connectivity_real.dart';
 import 'package:connectivity_plus_linux/src/network_manager.dart';
 import 'package:connectivity_plus_platform_interface/connectivity_plus_platform_interface.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 
-class MockNetworkManager extends Mock implements NetworkManager {}
+import 'connectivity_plus_linux_test.mocks.dart';
 
+@GenerateMocks([], customMocks: [
+  MockSpec<NetworkManager>(returnNullOnMissingStub: true),
+])
 void main() {
   test('wireless', () async {
     final linux = ConnectivityLinux();
