@@ -18,7 +18,7 @@ class NetworkInformationApiConnectivityPlugin extends ConnectivityPlugin {
 
   /// The constructor of the plugin.
   NetworkInformationApiConnectivityPlugin()
-      : this.withConnection(html.window.navigator.connection);
+      : this.withConnection(html.window.navigator.connection!);
 
   /// Creates the plugin, with an override of the NetworkInformation object.
   @visibleForTesting
@@ -32,8 +32,8 @@ class NetworkInformationApiConnectivityPlugin extends ConnectivityPlugin {
     return networkInformationToConnectivityResult(_networkInformation);
   }
 
-  StreamController<ConnectivityResult> _connectivityResultStreamController;
-  Stream<ConnectivityResult> _connectivityResultStream;
+  StreamController<ConnectivityResult>? _connectivityResultStreamController;
+  late Stream<ConnectivityResult> _connectivityResultStream;
 
   /// Returns a Stream of ConnectivityResults changes.
   @override
@@ -42,7 +42,7 @@ class NetworkInformationApiConnectivityPlugin extends ConnectivityPlugin {
       _connectivityResultStreamController =
           StreamController<ConnectivityResult>();
       setProperty(_networkInformation, 'onchange', allowInterop((_) {
-        _connectivityResultStreamController
+        _connectivityResultStreamController!
             .add(networkInformationToConnectivityResult(_networkInformation));
       }));
       // TODO: Implement the above with _networkInformation.onChange:
@@ -54,7 +54,7 @@ class NetworkInformationApiConnectivityPlugin extends ConnectivityPlugin {
       // onChange Stream upon hot restart.
       // https://github.com/dart-lang/sdk/issues/42679
       _connectivityResultStream =
-          _connectivityResultStreamController.stream.asBroadcastStream();
+          _connectivityResultStreamController!.stream.asBroadcastStream();
     }
     return _connectivityResultStream;
   }
